requirements.txt
pyyaml
pystray
pillow
psutil
pywin32
